ai_engine/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
google-generativeai
python-dotenv
python-multipart
openpyxl
httpx
